Input: iqs269a - add support for SYNC UI
This patch adds support for the device's SYNC UI function, which
allows the device to be started and stopped by a GPIO to inhibit
sensing while known sources of noise are active.

Drivers may leverage the SYNC UI by calling iqs269_sync() before
and after the noise source is enabled and disabled, respectively.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>