Input: iqs269a - configure device with a single block write
Unless it is being done as part of servicing a soft reset interrupt,
configuring channels on-the-fly (as is the case when writing to the
ati_trigger attribute) may cause GPIO3 (which reflects the state of
touch for a selected channel) to be inadvertently asserted.

To solve this problem, follow the vendor's recommendation and write
all channel configuration as well as the REDO_ATI register field as
part of a single block write. This ensures the device has been told
to re-calibrate itself following an I2C stop condition, after which
sensing resumes and GPIO3 may be asserted.

This change elicits minor re-shuffling of the driver's private data;
however the driver is ultimately made smaller now that back-to-back
register reads and writes have been replaced with single contiguous
operations.

To facilitate this change, some unnecessary validation of ch_num is
removed from iqs269_ati_*_set() and iqs269_ati_*_set() as ch_num is
already bounds-checked elsewhere.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>