Input: iqs269a - allow any channel to be a key or switch
The present implementation limits switches to the dedicated Hall
sensing channel, but some systems may call for any capacitive or
inductive sensing channel to be represented as a switch.

Therefore, borrow the 'linux,input-type' property from gpio-keys
such that any channel can be mapped to EV_KEY or EV_SW. Omitting
this property preserves the original behavior.

To facilitate this change, the 1-D 'keycode' array is renamed to
the 2-D 'kp_code' array. The nested loops in iqs269_input_init()
and iqs269_report() are made simpler as a result.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>