iio: proximity: iqs7219: Refresh system setup register cache
It has been observed that on other variants of silicon, the reserved
fields of the system setup register may change after other registers
are written.

As such, the read-modify-write operation which triggers ATI must use
the contents of the register as they appear following initialization
and not device reset.

This behavior has not been observed in the case of IQS7219A; however,
the register is read once more prior to being written as a matter of
best practice and out of an abundance of caution.

Fixes: 97fcf3fbefe8 ("iio: proximity: Add support for Azoteq IQS7219A")
Signed-off-by: Jeff LaBundy <jeff@labundy.com>