iio: proximity: iqs7219: Acknowledge reset before writing registers
If the device suffers a spurious reset while reacting to a previous
spurious reset, the second reset interrupt is preempted because the
ACK_RESET bit is written according to the order of the register map.

To solve this problem, write the ACK_RESET bit prior to writing any
other registers. This ensures that any registers written before the
second spurious reset will be rewritten.

Fixes: 97fcf3fbefe8 ("iio: proximity: Add support for Azoteq IQS7219A")
Signed-off-by: Jeff LaBundy <jeff@labundy.com>