Input: iqs9150 - add key event support
Add the ability to report a key event in response to each available
one and two finger gesture, the ALP proximity event, and the switch
input pin of the device. These events are grouped under an optional
"keyboard" input device.

As part of this change, add support for device tree properties that
correspond to gestures. Parsing of select ALP-related properties is
refactored as well.

Furthermore, the trackpad input device is now registered regardless
of the number of contacts specified; it is also registered first so
that its assigned number remains the same regardless of whether the
keyboard input device is present.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>