Input: iqs9150 - add support for VDD regulator
Add a means to request a regulator corresponding to the VDD supply
of the device. The regulator remains active for the device's whole
lifetime, and need not be controlled dynamically.

With this change, the device is placed in hardware reset following
driver detach, to ensure the GPIO connected to the MCLR input does
not backpower the VDD supply.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>