Input: iqs9150 - write communication timeout independently
Write the communication timeout register before starting any burst
reads or writes so that the timeout can be increased before it may
be reached during a lengthy transaction.

As a result, the maximum burst length can be set to a higher value
since the device is not the limiting factor so long as the timeout
is chosen appropriately.

Signed-off-by: Jeff LaBundy <jeff@labundy.com>