Input: iqs9150 - do not overwrite configuration register
Enabling key events overwrites changes made to the configuration
register (0x11BE) on behalf of the device tree because its value
is stored before the device tree is parsed.

Solve this problem by reversing the order of operations, thereby
preserving any custom fields in the configuration register.

As part of this change, skip any key events which do not specify
a type; this prevents the ALP event from being enabled if no key
code is defined.

Fixes: 2dc7f039ea5b ("Input: iqs9150 - add key event support")
Signed-off-by: Jeff LaBundy <jeff@labundy.com>