Input: iqs9150 - always terminate persistent gestures
Persistent gestures (e.g. press and hold or palm) do not generate
an interrupt upon release; rather, trailing interrupts associated
with trackpad movement are used to detect whether the gesture has
been terminated.

This method is not reliable, and can occasionally result in stuck
keys. Solve this problem by starting a watchdog when a persistent
gesture is detected. The watchdog is stopped if another interrupt
arrives shortly after, and all corresponding keys are released if
the watchdog bites.

As part of this change, ignore persistent gestures detected while
the trackpad is not in a state of touch; this avoids phantom keys
that may appear when the device recovers from a spurious hardware
reset event.

Fixes: 2dc7f039ea5b ("Input: iqs9150 - add key event support")
Signed-off-by: Jeff LaBundy <jeff@labundy.com>